Added In Ability to Edit Job Information

+++ app/src/main/java/micahsquad/com/worklogassistant/Job.java
@@ -4,7 +4,7 @@ package micahsquad.com.worklogassistant;
  * Created by Micah on 7/20/2016.
  */
 public class Job {
-    private int jobid;
+    private long jobid;
     private String name;
     private String position;
     private Double pay;
@@ -13,16 +13,16 @@ public class Job {
 
     }
 
-    public Job(int jobid, String name, String position, Double pay){
+    public Job(long jobid, String name, String position, Double pay){
         this.jobid = jobid;
         this.name = name;
         this.position = position;
         this.pay = pay;
     }
 
-    public int getJobId() { return jobid; }
+    public long getJobId() { return jobid; }
 
-    public void setJobId(int jobid) { this.jobid = jobid; }
+    public void setJobId(long jobid) { this.jobid = jobid; }
 
     public String getName(){
         return name;
